+++ src/main.cpp
@@ -13,7 +13,6 @@
 #include <unistd.h>
 
 #include "aot/aot.h"
-#include "ast/bpforc/bpforc.h"
 #include "ast/pass_manager.h"
 
 #include "ast/passes/codegen_llvm.h"
@@ -894,8 +893,6 @@ int main(int argc, char* argv[])
   BpfBytecode bytecode;
   try
   {
-    std::unique_ptr<BpfOrc> bpforc;
-
     llvm.generate_ir();
     if (bt_debug == DebugLevel::kFullDebug)
     {
@@ -932,16 +929,7 @@ int main(int argc, char* argv[])
       llvm.emit_elf(args.output_elf);
       return 0;
     }
-    bpforc = llvm.emit();
-    if (bt_debug == DebugLevel::kFullDebug)
-    {
-      std::cout << "\nLLVM JITDLib state\n";
-      std::cout << "------------------\n\n";
-      raw_os_ostream os(std::cout);
-      bpforc->dump(os);
-    }
-
-    bytecode = bpforc->getBytecode();
+    bytecode = std::move(llvm.emit());
   }
   catch (const std::system_error& ex)
   {