+++ src/map.cpp
@@ -126,7 +126,7 @@ Map::Map(const std::string &name,
 Map::Map(const SizedType &type) : IMap(type)
 {
   int key_size = 4;
-  int value_size = sizeof(uintptr_t) * type.stack_type.limit;
+  int value_size = sizeof(uint64_t) * type.stack_type.limit;
   int max_entries = 128 << 10;
   int flags = 0;
 
@@ -134,13 +134,19 @@ Map::Map(const SizedType &type) : IMap(type)
       map_type_, "stack", key_size, value_size, max_entries, flags);
   if (mapfd_ < 0)
   {
-    LOG(ERROR)
-        << "failed to create stack id map\n"
-        // TODO (mmarchini): Check perf_event_max_stack in the semantic_analyzer
-        << "This might have happened because kernel.perf_event_max_stack "
-        << "is smaller than " << type.stack_type.limit
-        << ". Try to tweak this value with "
-        << "sysctl kernel.perf_event_max_stack=<new value>";
+    LOG(ERROR) << "failed to create stack id map: " << strerror(errno);
+    if (errno == ENOMEM)
+      LOG(ERROR) << "Tried to allocate " << max_entries << " entries of size "
+                 << value_size
+                 << ". Consider limiting the number of stack frames captured.";
+    else
+      LOG(ERROR)
+          // TODO (mmarchini): Check perf_event_max_stack in the
+          // semantic_analyzer
+          << "This might have happened because kernel.perf_event_max_stack "
+          << "is smaller than " << type.stack_type.limit
+          << ". Try to tweak this value with "
+          << "sysctl kernel.perf_event_max_stack=<new value>";
   }
 }