+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to
 ## Unreleased
 
 #### Added
+- Support for 32-bit ARM systems
+  - [#2360](https://github.com/iovisor/bpftrace/pull/2360)
+#### Changed
+#### Deprecated
+#### Removed
+#### Fixed
+- Fix pointer/register loads on 32-bit architectures
+  - [#2361](https://github.com/iovisor/bpftrace/pull/2361)
+#### Docs
+#### Tools
+
+## [0.16.0] 2022-08-30
+
+#### Added
 - Add builtin: `numaid`
   - [#2177](https://github.com/iovisor/bpftrace/pull/2177)
 - Add helper verifier error handling
@@ -27,6 +41,8 @@ and this project adheres to
   - [#2264](https://github.com/iovisor/bpftrace/pull/2264)
 - Helper errors (-k, -kk options) are now emitted to text or json output
   - [#2326](https://github.com/iovisor/bpftrace/pull/2326)
+- kprobe offset verification is now optional, without requiring --unsafe
+  - [#2332](https://github.com/iovisor/bpftrace/pull/2332)
 
 #### Deprecated
 #### Removed
@@ -50,7 +66,6 @@ and this project adheres to
 - Fix access to ctx
   - [#2343](https://github.com/iovisor/bpftrace/pull/2343)
 
-#### Added
 #### Docs
 #### Tools
 - Add sslsnoop and ssllatency tools